improve pedantic compliance
	modified:   Makefile.common
	renamed:    source/csx_test_arm_asm.c -> archive/csx_test_arm_asm.c
	renamed:    source/soc_core_shifter.c -> archive/soc_core_shifter.c
	modified:   include/config.h
	modified:   include/csx_data.h
	modified:   include/csx_mmio.h
	modified:   include/csx_test_arm_inst.h
	modified:   include/csx_test_utility.h
	modified:   include/soc_core_decode.h
	modified:   include/soc_mmio.h
	modified:   source/csx_data.c
	modified:   source/csx_mem.c
	modified:   source/csx_mmio.c
	modified:   source/csx_soc.c
	modified:   source/csx_test.c
	modified:   source/csx_test_arm.c
	modified:   source/csx_test_arm_inst.c
	modified:   source/csx_test_thumb.c
	modified:   source/csx_test_thumb_asm.c
	modified:   source/csx_test_utility.c
	modified:   source/main.c
	modified:   source/soc_core_arm.c
	modified:   source/soc_core_thumb.c
	modified:   source/soc_mmio_cfg.c
	modified:   source/soc_mmio_dpll.c
	modified:   source/soc_mmu.c
	modified:   source/soc_omap_timer.c

+++ include/csx_test_utility.h
@@ -12,7 +12,7 @@ void _assert_cpsr_xpsr_mask(csx_test_p t, uint cpsr, uint xpsr, uint mask);
 void _assert_nzc(csx_test_p t, int n, int z, int c);
 void _assert_nzcv(csx_test_p t, int n, int z, int c, int v);
 void _cxx(csx_test_p t, uint32_t value, uint8_t size);
-uint32_t _test_value(uint8_t i);
+uint _test_value(uint i);
 
 /* **** */
